I have added a new resource to the repos.
I have added various resources in HTML, CSS, and Javascript. This resource provides a comprehensive overview of the features of HTML, which can be a great starting point for anyone new to HTML or looking to brush up on their existing knowledge and other resource provides an overview of the different types of CSS, which is a useful reference for anyone looking to learn more about the subject. The list of the best JavaScript books available and can provide great insight into the language.